Extract brush editor mutation helpers
